+++ sources/core/Ggio.BikeSherpa.Backend.Domain/DeliveryAggregate/Bootstrap.cs
@@ -17,7 +17,6 @@ public static class Bootstrap
                services.AddScoped<IPricingStrategy, SimpleDeliveryStrategy>();
                services.AddScoped<IPricingStrategy, TourDeliveryStrategy>();
                services.AddScoped<IDeliveryChangeTimeService, DeliveryChangeTimeService>();
-               services.AddScoped<IDeliveryStepAttachmentSaveService, DeliveryStepAttachmentSaveService>();
                return services;
           }
