+++ sources/core/Ggio.BikeSherpa.Backend.Infrastructure/Services/DeliveryStepAttachmentSaveService.cs
@@ -1,4 +1,4 @@
-namespace Ggio.BikeSherpa.Backend.Domain.DeliveryAggregate.Services.Step;
+namespace Ggio.BikeSherpa.Backend.Infrastructure.Services;
 
 public class DeliveryStepAttachmentSaveService: IDeliveryStepAttachmentSaveService
 {
